sophos connector renamed to SophosEP
