--- Solutions/Sophos/Data Connectors/AzureFunctionSophos/function.json
@@ -1,12 +1,0 @@
-{
-    "scriptFile": "main.py",
-    "bindings": [
-      {
-        "name": "mytimer",
-        "type": "timerTrigger",
-        "direction": "in",
-        "schedule": "0 */5 * * * *",
-        "runOnStartup": false
-      }
-    ]
-  }
