--- Solutions/Sophos/Data Connectors/host.json
@@ -1,15 +1,0 @@
-{
-    "version": "2.0",
-    "logging": {
-      "applicationInsights": {
-        "samplingSettings": {
-          "isEnabled": true,
-          "excludedTypes": "Request"
-        }
-      }
-    },
-    "extensionBundle": {
-      "id": "Microsoft.Azure.Functions.ExtensionBundle",
-      "version": "[1.*, 2.0.0)"
-    }
-  }
