--- Solutions/Sophos/Data Connectors/proxies.json
@@ -1,4 +1,0 @@
-{
-    "$schema": "http://json.schemastore.org/proxies",
-    "proxies": {}
-}
