--- Solutions/Sophos/Data Connectors/requirements.txt
@@ -1,3 +1,0 @@
-requests==2.25.1
-azure-storage-file-share==12.3.0
-azure-functions==1.6.0
