+++ Solutions/SophosEP/Data Connectors/AzureFunctionSophos/function.json
@@ -1,0 +1,12 @@
+{
+    "scriptFile": "main.py",
+    "bindings": [
+      {
+        "name": "mytimer",
+        "type": "timerTrigger",
+        "direction": "in",
+        "schedule": "0 */5 * * * *",
+        "runOnStartup": false
+      }
+    ]
+  }
